+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 0.1.3 / 2025-10-29
 
-* Added documentation for new options: `image`, `name`, `label`, `style`, `git_url`
+* Added documentation for new options:
+  `alt`, `badge_svg`, `class`, `dist_url`, `git_url`, `image`, `name`, `label`,
+  `style`, and `title`
 * Added `spec.platform` to `jekyll_badge.gemspec` because `RubyGems.org` now requires it
 * Long labels are now wrapped after underscores